// src/shared/generated/api/generated/api.ts
export interface WordExtendedInfo {
    /**
     * 
     * @type {string}
     * @memberof WordExtendedInfo
     */
    'word': string;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'start_millis': number;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'end_millis': number;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'accuracy': number;
    /**
     * 
     * @type {number}
     * @memberof WordExtendedInfo
     */
    'speaker_id'?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof WordExtendedInfo
     */
    'is_obscene'?: boolean;
}